pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coderace"
version = "0.1.0"
description = "Race coding agents against each other on real tasks"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "mikiships" }]
keywords = ["ai", "coding-agents", "benchmark", "claude", "codex", "aider"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
coderace = "coderace.cli:app"

[project.urls]
Homepage = "https://github.com/mikiships/coderace"
Repository = "https://github.com/mikiships/coderace"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-mock>=3.0", "ruff>=0.4.0"]
